chore: move retry helper to utils
